fix arrays of length 1
in the previous configuration the verilated class received a -1
in case the value of the single array entry was set to 1, because
it was converted to signed, and a signed integer with length 1
has only the values 0 and -1.

on the verilator side, the assignment of -1 to the port had the effect
that the the generated vdc file was corrupted at that point (maybe
this is a bug in verilator.